Refactor service setup with extension methods
Introduce new extension methods for setting up services in the application, including authentication, dependency injection, AutoMapper, and Swagger. These methods are encapsulated in static classes within the `MarCorp.DemoBack.Services.WebApi.Modules` namespace.

- `AuthenticationExtensions`: Adds JWT Bearer authentication configuration.
- `InjectionExtensions`: Sets up dependency injection for services, repositories, and logging.
- `MapperExtensions`: Configures AutoMapper services.
- `SwaggerExtensions`: Configures Swagger for API documentation.

Updated `Program.cs` to use these new extension methods, removing previous inline configurations. This change simplifies service setup and enhances code modularity and maintainability.
