+++ MarCorp.DemoBack.Services.WebApi/Program.cs
@@ -1,130 +1,17 @@
-using MarCorp.DemoBack.Application.Interface;
-using MarCorp.DemoBack.Application.Main;
-using MarCorp.DemoBack.Data.Connections;
-using MarCorp.DemoBack.Data.Interface;
-using MarCorp.DemoBack.Data.Repository;
-using MarCorp.DemoBack.Domain.Core;
-using MarCorp.DemoBack.Domain.Interface;
-using MarCorp.DemoBack.Services.WebApi.Helpers;
-using MarCorp.DemoBack.Support.Common;
-using MarCorp.DemoBack.Support.Logging;
-using MarCorp.DemoBack.Support.Mapper;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.OpenApi.Models;
-using System.Reflection;
-using System.Text;
+using MarCorp.DemoBack.Services.WebApi.Modules;
+using MarCorp.DemoBack.Services.WebApi.Modules.Authentication;
+using MarCorp.DemoBack.Services.WebApi.Modules.Injection;
+using MarCorp.DemoBack.Services.WebApi.Modules.Swagger;
 
 var builder = WebApplication.CreateBuilder(args);
+var configuration = new ConfigurationBuilder().SetBasePath(Directory.GetCurrentDirectory()).AddJsonFile("appsettings.json", optional: true, reloadOnChange: true).Build();
 
 builder.Services.AddControllers();
 builder.Services.AddOpenApi();
-
-var appSettingsSection = builder.Configuration.GetSection("Config");
-builder.Services.Configure<AppSettings>(appSettingsSection);
-
-var configuration = new ConfigurationBuilder().SetBasePath(Directory.GetCurrentDirectory()).AddJsonFile("appsettings.json", optional: true, reloadOnChange: true).Build();
-builder.Services.AddSingleton<IConfiguration>(configuration);
-
-builder.Services.AddSingleton<IConnectionFactory, ConnectionFactory>();
-builder.Services.AddScoped<ICustomerApplication, CustomerApplication>();
-builder.Services.AddScoped<ICustomerDomain, CustomerDomain>();
-builder.Services.AddScoped<ICustomerRepository, CustomerRepository>();
-builder.Services.AddScoped<IUsersRepository, UsersRepository>();
-builder.Services.AddScoped<IUsersApplication, UsersApplication>();
-builder.Services.AddScoped<IUsersDomain, UsersDomain>();
-builder.Services.AddAutoMapper(typeof(MappingsProfile));
-builder.Services.AddScoped(typeof(IAppLogger<>), typeof(LoggerAdapter<>));
-builder.Services.AddSwaggerGen(c =>
-{
-    c.SwaggerDoc("v1", new OpenApiInfo
-    {
-        Version = "v1",
-        Title = "MarCorp Technology Services API",
-        Description = "Example ASP.NET Core Web API",
-        TermsOfService = new Uri("https://example.com/terms"),
-        Contact = new OpenApiContact
-        {
-            Name = "Daniel Mar",
-            Email = "daniel.mar.iest.edu.mx",
-            Url = new Uri("https://example.com/contacto")
-        },
-        License = new OpenApiLicense
-        {
-            Name = "Use under LICX",
-            Url = new Uri("https://example.com/licencia")
-        }
-    });
-    // Set the comments path for the Swagger JSON and UI.
-    var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-    var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-    c.IncludeXmlComments(xmlPath);
-    c.AddSecurityDefinition("Authorization", new OpenApiSecurityScheme
-    {
-        Description = "Authorization by API key.",
-        In = ParameterLocation.Header,
-        Type = SecuritySchemeType.ApiKey,
-        Name = "Authorization"
-    });
-
-    c.AddSecurityRequirement(new OpenApiSecurityRequirement
-    {
-        {
-            new OpenApiSecurityScheme
-            {
-                Reference = new OpenApiReference
-                {
-                    Type = ReferenceType.SecurityScheme,
-                    Id = "Authorization"
-                }
-            },
-            new string[] { }
-        }
-    });
-});
-
-var appSettings = appSettingsSection.Get<AppSettings>();
-var key = Encoding.ASCII.GetBytes(appSettings.Secret);
-var Issuer = appSettings.Issuer;
-var Audience = appSettings.Audience;
-
-builder.Services.AddAuthentication(x =>
-{
-    x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-}).AddJwtBearer(x =>
-{
-    x.Events = new JwtBearerEvents
-    {
-        OnTokenValidated = context =>
-        {
-            var userId = int.Parse(context.Principal.Identity.Name);
-            return Task.CompletedTask;
-        },
-
-        OnAuthenticationFailed = context =>
-        {
-            if (context.Exception.GetType() == typeof(SecurityTokenExpiredException))
-            {
-                context.Response.Headers.Add("Token-Expired", "true");
-            }
-            return Task.CompletedTask;
-        }
-    };
-    x.RequireHttpsMetadata = false;
-    x.SaveToken = false;
-    x.TokenValidationParameters = new TokenValidationParameters
-    {
-        ValidateIssuerSigningKey = true,
-        IssuerSigningKey = new SymmetricSecurityKey(key),
-        ValidateIssuer = true,
-        ValidIssuer = Issuer,
-        ValidateAudience = true,
-        ValidAudience = Audience,
-        ValidateLifetime = true,
-        ClockSkew = TimeSpan.Zero
-    };
-});
+builder.Services.AddSwagger();
+builder.Services.AddAuthentication(configuration);
+builder.Services.AddMapper();
+builder.Services.AddInjection(configuration);
 
 var app = builder.Build();
